H-Deformable-DETR/requirements.txt
pycocotools==2.0.7
tqdm
cython
scipy
wandb
timm
einops
torchinfo
Flask>=2.2.2